Move sorting to HomeView
