+++ src/views/HomeView.vue
@@ -16,13 +16,15 @@
     :crns="crns"
     @companies-loaded="companiesLoaded" />
   <CompanyOutput
-    v-if="loadedCompanies.length > 0"
-    :companies="loadedCompanies" />
+    v-if="companies.length > 0"
+    :sorted-companies="sortedCompanies"
+    :sorting="sorting"
+    @sort="s => sorting = s" />
 </template>
 
 <script lang="ts">
 import { defineComponent } from 'vue';
-import Company from '@/models/Company';
+import Company, { CompanySorting } from '@/models/Company';
 import { ICompaniesHouseApi } from '@/models/CompaniesHouseApi';
 import ApiKeyInput from '@/components/ApiKeyInput.vue';
 import CompanyInput from '@/components/CompanyInput.vue';
@@ -43,11 +45,52 @@ export default defineComponent({
       companiesHouseApi: null as (ICompaniesHouseApi | null),
       crns: [] as string[],
       loading: false,
-      loadedCompanies: [] as Company[],
+      companies: [] as Company[],
+      sorting: null as CompanySorting | null,
     };
   },
+  computed: {
+    sortedCompanies(): Company[] {
+      const sorting = this.sorting;
+      if (sorting === null) return this.companies;
+
+      const sortedCompanies = structuredClone(this.companies);
+      sortedCompanies.sort((a: Company, b: Company) => {
+        const orderBy = sorting.order === 'asc' ? 1 : -1;
+        const valA = a[sorting.column];
+        const valB = b[sorting.column];
+        if (!valA) return orderBy * -1;
+        if (!valB) return orderBy * 1;
+
+        const sortDate = (dateA: string, dateB: string) => {
+          const sortableA = dateA.split('/').reverse().join('');
+          const sortableB = dateB.split('/').reverse().join('');
+          return sortableA.localeCompare(sortableB) * orderBy;
+        };
+
+        const sortStringArray = (arrayA: string[], arrayB: string[]) => {
+          const sortableA = arrayA.join('');
+          const sortableB = arrayB.join('');
+          return sortableA.localeCompare(sortableB) * orderBy;
+        };
+
+        switch (sorting.column) {
+          case 'confirmationStatementDue':
+          case 'accountsDue':
+            return sortDate(valA as string, valB as string);
+          case 'directors':
+            return sortStringArray(valA as string[], valB as string[]);
+          default:
+            if (valA > valB) return orderBy * 1;
+            if (valA < valB) return orderBy * -1;
+            return 1;
+        }
+      });
+      return sortedCompanies;
+    },
+  },
   beforeRouteLeave() {
-    if (this.loadedCompanies.length > 0) {
+    if (this.companies.length > 0) {
       const answer = window.confirm('Are you sure? All loaded data will be lost.');
       if (!answer) return false;
       window.removeEventListener('beforeunload', this.beforeUnloadHandler);
@@ -60,8 +103,8 @@ export default defineComponent({
       return (e.returnValue = 'Are you sure? All loaded data will be lost.');
     },
     companiesLoaded(companies: { companies: Company[] }) {
-      this.loadedCompanies = companies.companies;
-      if (this.loadedCompanies.length > 0) {
+      this.companies = companies.companies;
+      if (this.companies.length > 0) {
         window.addEventListener('beforeunload', this.beforeUnloadHandler);
       } else {
         window.removeEventListener('beforeunload', this.beforeUnloadHandler);
